+++ yarss2/__init__.py
@@ -10,7 +10,7 @@
 # Copyright (C) 2007-2009 Andrew Resch <andrewresch@gmail.com>
 # Copyright (C) 2009 Damien Churchill <damoxc@gmail.com>
 #
-# v2.2.0 maintenance release and WebUI port:
+# v2.2.x maintenance release and WebUI port:
 # Copyright (C) 2026 Sam Mahdi
 #
 # This file is part of YaRSS2 and is licensed under GNU General Public License 3.0, or later, with